+++ app/src/main/java/woowacourse/paint/ColorDiffUtilCallback.kt
@@ -1,6 +1,7 @@
 package woowacourse.paint
 
 import androidx.recyclerview.widget.DiffUtil
+import woowacourse.paint.model.ColorBox
 
 object ColorDiffUtilCallback : DiffUtil.ItemCallback<ColorBox>() {
     override fun areItemsTheSame(oldItem: ColorBox, newItem: ColorBox): Boolean {
