+++ app/src/main/java/woowacourse/paint/ColorViewHolder.kt
@@ -4,6 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.paint.databinding.ItemColorBinding
+import woowacourse.paint.model.ColorBox
 
 class ColorViewHolder private constructor(
     private val binding: ItemColorBinding,
