+++ app/src/main/java/woowacourse/paint/DrawingCanvas.kt
@@ -8,6 +8,8 @@ import android.graphics.Paint
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
+import woowacourse.paint.model.DrawingElement
+import woowacourse.paint.model.DrawingHistory
 
 class DrawingCanvas @JvmOverloads constructor(
     context: Context,
