+++ app/src/main/java/woowacourse/paint/OnColorClickListener.kt
@@ -1,5 +1,7 @@
 package woowacourse.paint
 
+import woowacourse.paint.model.ColorBox
+
 interface OnColorClickListener {
     fun onColorClick(colorBox: ColorBox)
 }
