+++ app/src/main/java/woowacourse/paint/model/ColorBox.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.model
 
 data class ColorBox(
     val color: Int,
